app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.6",
    "pinecone>=7.3.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]
